Move MaterializeEncodingInfo to Codegen/Dialect/Utils/
