+++ compiler/src/iree/compiler/Codegen/Dialect/Codegen/Utils/Utils.h
@@ -8,11 +8,17 @@
 #define IREE_COMPILER__CODEGEN_DIALECT_CODEGEN_UTILS_H_
 
 #include <cstdint>
+
 #include "llvm-c/TargetMachine.h"
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/Support/raw_ostream.h"
+#include "mlir/Support/LLVM.h"
 
 namespace mlir::iree_compiler::IREE::Codegen {
+
+//===----------------------------------------------------------------------===//
+// Layout Structs.
+//===----------------------------------------------------------------------===//
 
 // Metadata for a swizzle, that is, an (expand_shape -> transposition)
 // pair of ops performing a change of layout within the tiles. This is used
@@ -80,6 +86,27 @@ llvm::raw_ostream &operator<<(llvm::raw_ostream &os, TileSwizzle::Dim dim);
 
 llvm::raw_ostream &operator<<(llvm::raw_ostream &os,
                               const TileSwizzle &swizzle);
+
+/// Container of information needed to materialize the layout transformations.
+struct MaterializeEncodingInfo {
+  // The next 3 fields are used to create a `tensor.pack` or `tensor.unpack` op,
+  // changing the overall layout between row-major and tiled (where each tile is
+  // row-major).
+  SmallVector<int64_t> innerDimsPos;
+  SmallVector<int64_t> innerTileSizes;
+  SmallVector<int64_t> outerDimsPerm;
+
+  // The optional swizzle, see the comment on TileSwizzle. Only used on GPU.
+  std::optional<TileSwizzle> swizzle;
+};
+
+//===----------------------------------------------------------------------===//
+// Layout Utilities.
+//===----------------------------------------------------------------------===//
+
+/// Concatenates the vectors.
+SmallVector<int64_t>
+getExpandedTileShape(const TileSwizzle::ExpandShapeType &expandShape);
 
 } // namespace mlir::iree_compiler::IREE::Codegen
